refactor: restructure Search module
Search now has two subpackages: api & impl. Code is currently in impl.
