+++ feature/search/impl/src/main/res/mipmap-hdpi/ic_launcher.webp
