+++ feature/search/impl/src/main/res/mipmap-xhdpi/ic_launcher.webp
