+++ settings.gradle.kts
@@ -43,7 +43,8 @@ include(":feature:subreddit")
 include(":core:common")
 include(":feature:subredditwiki")
 include(":feature:user")
-include(":feature:search")
+include(":feature:search:impl")
+include(":feature:search:api")
 include(":core:markdown")
 include(":feature:video")
 include(":core:domain")
